Extract error tooltip directive from proofreadings controller
