+++ src/scripts/app/play/proofreadings/errorTooltip/directive.js
@@ -1,0 +1,9 @@
+'use strict';
+
+module.exports = function () {
+  return {
+    restrict: 'E',
+    templateUrl: 'error-tooltip.html',
+    controller: require('./controller.js')
+  };
+};
