+++ src/scripts/app/play/proofreadings/errorTooltip/error-tooltip.html
@@ -1,0 +1,20 @@
+<div ng-class="getErrorTooltipClass(index)" ng-show="!word.isBr() && word.hasErrorToShow()"
+  ng-style="word.tooltip.style">
+  <div ng-class="{incorrectError: word.hasIncorrectError(), notNecessaryError: word.hasNotNecessaryError(), correct: word.hasCorrect()}">
+    <div ng-class="getErrorTooltipTopClass(answerImageName(word.type))">
+      <div class="img-div">
+        <img ng-src="/assets/images/{{answerImageName(word.type)}}_icon.png" alt="{{word.type}}"></img>
+      </div>
+      <h4>{{word.type}}</h4>
+      <p class="counter">{{word.errorCounter()}}</p>
+    </div>
+    <div class="bottom-panel">
+      <h4>Correct Edit:</h4>
+      <p ng-show="word.plus">{{word.plus}}</p>
+      <p ng-show="!word.plus">{{word.text}}</p>
+      <h4 ng-show="!word.hasNotNecessaryError()">Grammatical Concept:</h4>
+      <p ng-show="!word.hasNotNecessaryError()">{{proofreadingPassage.getGrammaticalConcept(word)}}</p>
+      <button ng-click="word.nextAction.fn()">{{word.nextAction.title}} ➞</button>
+    </div>
+  </div>
+</div>
