+++ src/scripts/app/play/proofreadings/module.js
@@ -32,4 +32,5 @@ module.exports = angular.module('quill-grammar.play.proofreadings', [
   };
 })
 .directive('ngSize', require('./ngSize.js'))
+.directive('proofreadingErrorTooltip', require('./errorTooltip/directive.js'))
 .controller('ProofreadingPlayCtrl', require('./controller.js'));
