+++ src/scripts/app/play/proofreadings/passage.html
@@ -6,7 +6,7 @@
       ng-show="!word.isBr()"
       type="text"
       ng-class="{focused: word.isFocused, incorrectError: word.hasIncorrectError(), notNecessaryError: word.hasNotNecessaryError(), changed: word.text !== word.responseText, correct: word.hasCorrect(), underlined: needsUnderlining(word), tooltipped: word.tooltip.style}"
-      ng-readonly="passage.submitted"
+      ng-readonly="proofreadingPassage.submitted"
       ng-focus="word.isFocused = true"
       ng-change="proofreadingPassage.onInputChange(word)"
       ng-blur="word.isFocused = false"
@@ -14,25 +14,6 @@
       ng-trim="false"
       ng-model="word.responseText">
     </input>
-    <div ng-class="getErrorTooltipClass(index)" ng-show="!word.isBr() && word.hasErrorToShow()"
-      ng-style="word.tooltip.style">
-      <div ng-class="{incorrectError: word.hasIncorrectError(), notNecessaryError: word.hasNotNecessaryError(), correct: word.hasCorrect()}">
-        <div ng-class="getErrorTooltipTopClass(answerImageName(word.type))">
-          <div class="img-div">
-            <img ng-src="/assets/images/{{answerImageName(word.type)}}_icon.png" alt="{{word.type}}"></img>
-          </div>
-          <h4>{{word.type}}</h4>
-          <p class="counter">{{word.errorCounter()}}</p>
-        </div>
-        <div class="bottom-panel">
-          <h4>Correct Edit:</h4>
-          <p ng-show="word.plus">{{word.plus}}</p>
-          <p ng-show="!word.plus">{{word.text}}</p>
-          <h4 ng-show="!word.hasNotNecessaryError()">Grammatical Concept:</h4>
-          <p ng-show="!word.hasNotNecessaryError()">{{proofreadingPassage.getGrammaticalConcept(word)}}</p>
-          <button ng-click="word.nextAction.fn()">{{word.nextAction.title}} ➞</button>
-        </div>
-      </div>
-    </div>
+    <proofreading-error-tooltip></proofreading-error-tooltip>
   </div>
 </div>
